requirements.txt
asgiref==3.2.6
Babel==2.8.0
cffi==1.14.0
cryptography==2.8
Django==3.0.4
django-cors-headers==3.2.1
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
phonenumberslite==8.12.0
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
